chore: Refactor hook files
