+++ frontend/src/components/organisms/LatestUpdates.tsx
@@ -6,11 +6,11 @@ import { ShowMoreButton, Typography } from "@atoms";
 import { useTranslations } from "next-intl";
 import { VotesTable } from "./VotesTable";
 import { NotFound } from "./NotFound";
-import { countSelectedFilters, isEmpty, useManageQueryParams } from "@utils";
+import { countSelectedFilters, isEmpty } from "@utils";
 import { DataActionsBar } from "../molecules";
 import { LATEST_UPDATES_FILTERS, LATEST_UPDATES_SORTING } from "@consts";
 import { PaginationMeta, VotesTableI } from "@/lib/requests";
-import { usePagination } from "@/lib/utils/usePagination";
+import { usePagination, useManageQueryParams } from "@hooks";
 import { getLatestUpdates } from "@/lib/api";
 import { OpenPreviewReasoningModal } from "./types";
 import { useModal } from "@/context";
