+++ frontend/src/lib/hooks/useDocumentVisibility.tsx
