+++ frontend/src/lib/utils/index.ts
@@ -9,8 +9,3 @@ export * from "./getGovActionId";
 export * from "./string";
 export * from "./getProposalTypeLabel";
 export * from "./helpers";
-export * from "./useDebounce";
-export * from "./useManageQueryParams";
-export * from "./routes";
-export * from "./useDocumentVisibility";
-export * from "./useScreenDimension";
